Updated SSP Scenario Button/Feature and Minor Edits
Created the feature that allows access to an external CSV with a set of inputs. This also requires the CSV file to initialise 81 variables (that are hidden), and hiding the display of 81 vars in the summary table.

Minor unrelated edits include (1) Model: changing the custom diet so it is by kcal instead of % (2) Small fixes to the toolkit